Refactor FileReview
