+++ crates/types/src/lib.rs
@@ -14,17 +14,6 @@ pub use self::source_item::{
     MatchedItem, SourceItem,
 };
 
-/// The preview content is usually part of a file.
-#[derive(Clone, Debug)]
-pub struct PreviewInfo {
-    pub start: usize,
-    pub end: usize,
-    /// Line number of the line that should be highlighed in the preview window.
-    pub highlight_lnum: usize,
-    /// [start, end] of the source file.
-    pub lines: Vec<String>,
-}
-
 #[derive(Clone, Copy, Debug, Default)]
 pub enum CaseMatching {
     Ignore,
